# expression: implement vectorized evaluation for `builtinToSecondsSig`